I have not committed in far, far too long. I have gotten Swagger working, and done significant refactoring to reduce code duplication. I have also added a number of tests to the project for these new classes, but they are not yet complete

+++ src/main/java/com/doug/projects/transitdelayservice/entity/dynamodb/RouteTimestamp.java
@@ -21,7 +21,7 @@ public class RouteTimestamp {
     /**
      * The timestamp, gathered from GTFS realtime feed. DO NOT USE SYSTEM TIMESTAMP to avoid duplication
      */
-    public Integer timestamp;
+    public Long timestamp;
     /**
      * The list of routes, in the format of <code>BusStates</code> toString method. delay%closestStopId%tripId
      */
@@ -37,7 +37,7 @@ public class RouteTimestamp {
     }
 
     @DynamoDbSortKey
-    public Integer getTimestamp() {
+    public Long getTimestamp() {
         return timestamp;
     }
 }
